app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.5.2
    # via django
attrs==22.1.0
    # via jsonschema
build==0.9.0
    # via pip-tools
certifi==2022.12.7
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via pip-tools
cryptography==38.0.4
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
dj-rest-auth[with_social]==2.2.5
    # via -r requirements.in
django==4.1.4
    # via
    #   -r requirements.in
    #   dj-rest-auth
    #   django-allauth
    #   django-cors-headers
    #   djangorestframework
    #   drf-spectacular
django-allauth==0.50.0
    # via dj-rest-auth
django-cors-headers==3.13.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   dj-rest-auth
    #   drf-spectacular
drf-spectacular==0.24.2
    # via -r requirements.in
idna==3.4
    # via requests
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.17.3
    # via drf-spectacular
oauthlib==3.2.2
    # via requests-oauthlib
packaging==22.0
    # via build
pep517==0.13.0
    # via build
pip-tools==6.11.0
    # via -r requirements.in
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.6.0
    # via django-allauth
pyrsistent==0.19.2
    # via jsonschema
python3-openid==3.2.0
    # via django-allauth
pytz==2022.6
    # via djangorestframework
pyyaml==6.0
    # via drf-spectacular
requests==2.28.1
    # via
    #   django-allauth
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via django-allauth
sqlparse==0.4.3
    # via django
tomli==2.0.1
    # via
    #   build
    #   pep517
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.13
    # via requests
wheel==0.38.4
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
